+++ src/main.tsx
@@ -1,10 +1,14 @@
 import React from 'react'
 import ReactDOM from 'react-dom/client'
-import App from './App.tsx'
-import './index.css'
+import App from './tsx/App.tsx'
+import Footer from './tsx/layout/Footer.tsx'
+import Header from './tsx/layout/Header.tsx'
+import './css/site.css'
 
 ReactDOM.createRoot(document.getElementById('root')!).render(
-  <React.StrictMode>
-    <App />
-  </React.StrictMode>,
+    <React.StrictMode>
+        <Header />
+            <App />
+        <Footer />
+    </React.StrictMode>,
 )