+++ CHANGELOG.md
@@ -7,19 +7,39 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [v1.0.0] - 2024-07-06
+
+### Added
+
+- feat: update readme (#25)
+- feat: add stakeholder section as slider (#26)
+- feat: add introduction section (#30)
+- feat: add favicons (#31)
+- feat: add current process section (#36)
+- feat: add project page (#41)
+- feat: add dashboard preview section (#43)
+- feat: add advantages section (#45)
+- feat: add contact section (#46)
+- feat: add homepage hero (#50)
+- feat: add contact page (#53)
+- feat: add not found page (#54)
+- feat: add hero homepage overlays (#56)
+- feat: add background to sections (#59)
+
+### Changed
+
+-   Improve CI/CD pipeline for deployment, staging and production
+
 ## [v0.0.1] - 2024-05-26
 
 ### Changed
 
 -   Improve CI/CD pipeline
-
-## [0.0.0] - 2024-05-22
-
-### Added
-
 -   Initial release
 -   Create "Page under construction" Page
 
-[Unreleased]: https://github.com/SmartCityFlensburg/project-website/compare/v0.0.1...HEAD
+[Unreleased]: https://github.com/SmartCityFlensburg/project-website/compare/v1.0.0...HEAD
 
-[v0.0.1]: https://github.com/SmartCityFlensburg/project-website/compare/472d9c0...v0.0.1
+[v1.0.0]: https://github.com/SmartCityFlensburg/project-website/compare/v0.0.1...v1.0.0
+
+[v0.0.1]: https://github.com/SmartCityFlensburg/project-website/releases/tag/v0.0.1